backend/requirements.txt
# Django and core dependencies
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
python-decouple>=3.8
mysqlclient>=2.2.0

# File handling and media
Pillow>=10.0.0

# AI/ML libraries for personality analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
tensorflow>=2.19.0
keras>=3.0.0

# Audio processing
librosa>=0.10.0
pydub>=0.25.0
ffmpeg-python>=0.2.0
resampy>=0.4.0

# Computer vision for facial analysis
opencv-python>=4.8.0

# Additional packages (add these if needed later)
# nltk>=3.8.0
# textblob>=0.17.0
# spacy>=3.6.0
# celery>=5.3.0
# redis>=4.5.0
# psycopg2-binary>=2.9.0
# pytest>=7.4.0
# pytest-django>=4.5.0